==> 5a CTS prep.qmd <==
---
title: "5a Case Time Series Prep"
author: "Andrea Portt"
format: html
editor: visual
bibliography: references.bib
---

# Packages & data

```{r packages, message=FALSE}
library(dlnm) ; library(gnm) ; 
library(data.table) ; 
library(splines)
library(dplyr); 
library("stringr"); 
library(tidyr); 
library('correlation')
library(ggplot2)
library(tidyverse)

```

```{r}
#MB dataset 
load("~/2024_11_15_MB.Rdata")

```

# Correlation co-efficients

```{r}


#Calculate individual-level correlation coefficients
table <- data %>% 
        group_by(random_uid) %>% 
         correlation(select = c("no2","o3", "pm25", "so2", "temp_12", "humid_12000", "pres_change_24h"))           
table <- as.data.frame(table)

#correlation coefficients pooled for 1000 participants
data_group <- select(data, c(no2, o3, pm25, so2, temp_12, humid_12000, pres_change_24h))
data_group <- data_group[,2:8]

group_corr <- data_group %>% 
         correlation(select = c("no2","o3", "pm25", "so2", "temp_12", "humid_12000", "pres_change_24h"))           
group_corr <- as.data.frame(group_corr)

write.csv(group_corr, "group_corr.csv")


```

```{r}

table2 <- table %>%
  unite("parameter_joined", Parameter1, Parameter2, sep = "_")

table2 <- table2[1:3] #Remove unneccesary columns so that the spread can be one row per participant

summary_data <- table2 %>%
  spread(key = parameter_joined, value = r)

```

### Summarize correlations

```{r}

summary_data <- as.data.frame(summary_data)
summary_data$individual <- 1:1000

 #1 #humid_12 pressure change
q_humid_12000_pres_change_24h <- quantile(summary_data$`humid_12000_pres_change_24h`)
q_humid_12000_pres_change_24h <- as.data.frame(q_humid_12000_pres_change_24h)
q_humid_12000_pres_change_24h$quantiles <- c(0, 25, 50, 75, 100)

#2 no2 humid 12000
q_no2_humid_12000 <- quantile(summary_data$no2_humid_12000)
q_no2_humid_12000 <- as.data.frame(q_no2_humid_12000)
q_no2_humid_12000$quantiles <- c(0, 25, 50, 75, 100)

#3  "no2_o3"                                       
q_no2_o3 <- quantile(summary_data$no2_o3)
q_no2_o3 <- as.data.frame(q_no2_o3)
q_no2_o3$quantiles <- c(0, 25, 50, 75, 100)

#4 "no2_pm25"                                                                               
q_no2_pm25 <- quantile(summary_data$no2_pm25)
q_no2_pm25 <- as.data.frame(q_no2_pm25)
q_no2_pm25$quantiles <- c(0, 25, 50, 75, 100)

#5 "no2_pres_change_24h"
q_no2_pres_change_24h <- quantile(summary_data$no2_pres_change_24h)
q_no2_pres_change_24h <- as.data.frame(q_no2_pres_change_24h)
q_no2_pres_change_24h$quantiles <- c(0, 25, 50, 75, 100)

#6  "no2_so2" 
q_no2_so2 <- quantile(summary_data$no2_so2)
q_no2_so2 <- as.data.frame(q_no2_so2)
q_no2_so2$quantiles <- c(0, 25, 50, 75, 100)

#7 "no2_temp_12" 
q_no2_temp_12 <- quantile(summary_data$no2_temp_12)
q_no2_temp_12 <- as.data.frame(q_no2_temp_12)
q_no2_temp_12$quantiles <- c(0, 25, 50, 75, 100)

#8 "o3_humid_12000"
q_o3_humid_12000 <- quantile(summary_data$o3_humid_12000)
q_o3_humid_12000 <- as.data.frame(q_o3_humid_12000)
q_o3_humid_12000$quantiles <- c(0, 25, 50, 75, 100)

#9 "o3_pm25"  
q_o3_pm25 <- quantile(summary_data$o3_pm25)
q_o3_pm25 <- as.data.frame(q_o3_pm25)
q_o3_pm25$quantiles <- c(0, 25, 50, 75, 100)

#10 [11] "o3_pres_change_24h"                             
q_o3_pres_change_24h <- quantile(summary_data$o3_pres_change_24h)
q_o3_pres_change_24h <- as.data.frame(q_o3_pres_change_24h)
q_o3_pres_change_24h$quantiles <- c(0, 25, 50, 75, 100)

#11 "o3_so2"   
q_o3_so2 <- quantile(summary_data$o3_so2)
q_o3_so2 <- as.data.frame(q_o3_so2)
q_o3_so2$quantiles <- c(0, 25, 50, 75, 100)

#12 "o3_temp_12"
q_o3_temp_12 <- quantile(summary_data$o3_temp_12)
q_o3_temp_12 <- as.data.frame(q_o3_temp_12)
q_o3_temp_12$quantiles <- c(0, 25, 50, 75, 100)

#13 "pm25_humid_12000"
q_pm25_humid_12000 <- quantile(summary_data$pm25_humid_12000)
q_pm25_humid_12000 <- as.data.frame(q_pm25_humid_12000)
q_pm25_humid_12000$quantiles <- c(0, 25, 50, 75, 100)

#14 "pm25_pres_change_24h"
q_pm25_pres_change_24h <- quantile(summary_data$pm25_pres_change_24h)
q_pm25_pres_change_24h <- as.data.frame(q_pm25_pres_change_24h)
q_pm25_pres_change_24h$quantiles <- c(0, 25, 50, 75, 100)

#15 "pm25_so2"
q_pm25_so2 <- quantile(summary_data$pm25_so2)
q_pm25_so2 <- as.data.frame(q_pm25_so2)
q_pm25_so2$quantiles <- c(0, 25, 50, 75, 100)

#16  "pm25_temp_12"               
q_pm25_temp_12 <- quantile(summary_data$pm25_temp_12)
q_pm25_temp_12 <- as.data.frame(q_pm25_temp_12)
q_pm25_temp_12$quantiles <- c(0, 25, 50, 75, 100)

#17 "so2_humid_12000" 
q_so2_humid_12000 <- quantile(summary_data$so2_humid_12000)
q_so2_humid_12000 <- as.data.frame(q_so2_humid_12000)
q_so2_humid_12000$quantiles <- c(0, 25, 50, 75, 100)

#18 "so2_pres_change_24h" 
q_so2_pres_change_24h <- quantile(summary_data$so2_pres_change_24h)
q_so2_pres_change_24h <- as.data.frame(q_so2_pres_change_24h)
q_so2_pres_change_24h$quantiles <- c(0, 25, 50, 75, 100)

#19  "so2_temp_12" 
q_so2_temp_12 <- quantile(summary_data$so2_temp_12)
q_so2_temp_12 <- as.data.frame(q_so2_temp_12)
q_so2_temp_12$quantiles <- c(0, 25, 50, 75, 100)

#20 "temp_12_humid_12000"
q_temp_12_humid_12000 <- quantile(summary_data$temp_12_humid_12000)
q_temp_12_humid_12000 <- as.data.frame(q_temp_12_humid_12000)
q_temp_12_humid_12000$quantiles <- c(0, 25, 50, 75, 100)

#21 "temp_12_pres_change_24h" 
q_temp_12_pres_change_24h <- quantile(summary_data$temp_12_pres_change_24h)
q_temp_12_pres_change_24h <- as.data.frame(q_temp_12_pres_change_24h)
q_temp_12_pres_change_24h$quantiles <- c(0, 25, 50, 75, 100)


#record_sum$labels = factor(record_sum$labels, levels = c("0", "1-4", "5-10", "11-30", "31-50", "51+")) To make quantiles % but keep in order

#put all data frames into list
df_list <- list(q_humid_12000_pres_change_24h, q_no2_humid_12000,
               q_no2_o3, q_no2_pm25,
                q_no2_pres_change_24h, q_no2_so2,
                 q_no2_temp_12, q_o3_humid_12000, 
                q_o3_pm25, q_o3_pres_change_24h, 
                q_o3_so2, q_o3_temp_12, 
                q_pm25_humid_12000, q_pm25_pres_change_24h, 
                q_pm25_so2, q_pm25_temp_12, 
                q_so2_humid_12000, q_so2_pres_change_24h, 
                q_so2_temp_12,      q_temp_12_humid_12000, 
                q_temp_12_pres_change_24h)


#merge all data frames in list
variable_correlations <- df_list %>% 
                  reduce(full_join, by='quantiles')

variable_correlations <- variable_correlations %>% 
                        relocate(quantiles, .before = q_humid_12000_pres_change_24h)


#Subset to talk to Peter
chat_individual <- select(variable_correlations, c(quantiles, q_no2_o3, q_no2_pm25, q_no2_so2, q_no2_temp_12, q_temp_12_humid_12000))

#no2 <- ggplot(table_no2, aes(individual, `50%`, ymin = `25%`, ymax = `75%`)) +
 #           geom_pointrange() +
  #          labs( x = "Individual", y = expression("Range of NO"[2]* " IQRs"), subtitle =  expression('NO'[2]* 'Interquartile ranges of a random subset of 1000 participants'))
#no2



```

# Individual total ranges

3.  Check that ranges for individuals are wide enough to expect health impact if there is an association (Refer to review for ranges)

Confirm that there’s sufficient exposure variation *within* individuals to rationalize using fixed effects model.

Especially O3 - is it reasonable to model or are exposure ranges too tight?

```{r}

data$pm25 <- as.numeric(data$pm25)

#width by max - min
#histogram vs scatterplot of min and max ... error bars
#histogram doesn't tell starting point, so a scatterplot might be helpful
# y = width , x as starting point
# or y as uid, x as pm25 w max to min bars
#ggplot cheat sheet
#stdev per person could also be informative


range(data$pm25)


table_no2 <- data %>% 
         group_by(random_uid) %>% 
        summarize(min = min(no2), max = max(no2))
table_no2$mean <- rowMeans(table_no2[,c('min', 'max')], na.rm=TRUE)
table_no2$individual <- 1:1000


table_o3 <- data %>% 
         group_by(random_uid) %>% 
        summarize(min = min(o3), max = max(o3))
table_o3$mean <- rowMeans(table_o3[,c('min', 'max')], na.rm=TRUE)
table_o3$individual <- 1:1000

table_pm25 <- data %>% 
         group_by(random_uid) %>% 
        summarize(min = min(pm25), max = max(pm25))
table_pm25$mean <- rowMeans(table_pm25[,c('min', 'max')], na.rm=TRUE)
table_pm25$individual <- 1:1000


table_so2 <- data %>% 
         group_by(random_uid) %>% 
        summarize(min = min(so2), max = max(so2))
table_so2$mean <- rowMeans(table_so2[,c('min', 'max')], na.rm=TRUE)
table_so2$individual <- 1:1000

table_temp <- data %>% 
         group_by(random_uid) %>% 
        summarize(min = min(temp_12), max = max(temp_12))
table_temp$mean <- rowMeans(table_temp[,c('min', 'max')], na.rm=TRUE)
table_temp$individual <- 1:1000

table_humid <- data %>% 
         group_by(random_uid) %>% 
        summarize(min = min(humid_12), max = max(humid_12))
table_humid$mean <- rowMeans(table_humid[,c('min', 'max')], na.rm=TRUE)
table_humid$individual <- 1:1000

table_pressure <- data %>% 
                 group_by(random_uid) %>% 
                  summarize(min = min(pres_change_24h), max = max(pres_change_24h))
table_pressure$mean <- rowMeans(table_pressure[,c('min', 'max')], na.rm=TRUE)
table_pressure$individual <- 1:1000



```

# Plot individual exposure total ranges

```{r}

summary(MB$no2)

no2 <- ggplot(table_no2, aes(individual, mean, ymin = min, ymax = max)) +
            geom_pointrange() +
            labs( x = "Individual", y = "Range of NO2 exposure estimates", subtitle =  "NO2 ranges of a random subset of 1000 participants")
no2

 summary(MB$no2)
#   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 0.0085  2.4814  5.2152  7.2356  9.9326 60.3680 
 
 table_no2$width <- (table_no2$max - table_no2$min)

## Look at individual ranges of exposure
summary(table_no2$width)
 # Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
  # 1.981  23.224  33.613  33.857  43.615  60.126


o3 <- ggplot(table_o3, aes(individual, mean, ymin = min, ymax = max)) +
            geom_pointrange() +
            labs( x = "Individual", y = "Range of o3 exposure estimates", subtitle =  "o3 ranges of a random subset of 1000 participants")
o3

 summary(MB$o3)
#   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
 #  0.77   21.28   27.54   27.98   34.00  103.13 
 #O3 IQR is tight


pm25 <- ggplot(table_pm25, aes(individual, mean, ymin = min, ymax = max)) +
            geom_pointrange() +
            labs( x = "Individual", y = "Range of pm25 exposure estimates", subtitle =  "pm25 ranges of a random subset of 1000 participants")
pm25

 summary(MB$pm25)
#   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
 # 0.001   2.778   5.556   7.328   9.979  68.462 
 #pm25 is also tight


so2 <- ggplot(table_so2, aes(individual, mean, ymin = min, ymax = max)) +
            geom_pointrange() +
            labs( x = "Individual", y = "Range of so2 exposure estimates", subtitle =  "so2 ranges of a random subset of 1000 participants") #+
            #coord_cartesian( ylim = c(0, 10))
so2

#Highest readings for so2 are crazy-high
#And most of the ranges are around 0 - 5
#What ranges of SO2 were in my review? ~ 0.6 to 6 ppb IQRs. 
#Less than 5 ppb IQR probably isn't worth including
#Calculate IQR 

summary(MB$so2)
#    Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
 # 0.0000   0.4713   1.0529   1.8758   2.2278 288.9383 

#Will exclude SO2 from my analysis. Levels are too low to expect a measureable heath impact in this context. 
```

# Individual interquartile ranges

What's important is actually variation within specific months on given days for individuals.

Pick out Jan, April, June, Sept and plot ranges, or ranges on given days of week.

## NO2 - D.O.W.

This NO2 range with relation to results in review:

```{r}

#NO2
table_no2 <- do.call("rbind",tapply(data$no2, data$random_uid, quantile))
table_no2 <- as.data.frame(table_no2)
table_no2$individual <- 1:1000 #giving individuals a simple number for plotting

quantile(table_no2$`25%`)
#  0%       25%       50%       75%      100% 
# 0.116625  1.702794  3.351875  5.138962 11.418625 

quantile(table_no2$`75%`)
 #     0%      25%      50%      75%     100% 
 #0.37475  4.74120  9.26995 12.98660 23.08042 

no2 <- ggplot(table_no2, aes(individual, `50%`, ymin = `25%`, ymax = `75%`)) +
            geom_pointrange() +
            labs( x = "Individual", y = expression("Range of NO"[2]* " IQRs"), subtitle =  expression('NO'[2]* 'Interquartile ranges of a random subset of 1000 participants'))
no2

#As Peter pointed out, variation within specific day of week of given months are what's important because that's what we're comparing the onset day to. 

#So, subset by month and then facet by day of week. 

#Give individuals a simpler unique ID
data <- data %>% 
        group_by(random_uid) %>% 
        mutate(ID = cur_group_id())

#dataset of just Sundays in Dec 2017
Dec_2017_sundays <- data[data$year == 2017 &
                data$month == 12 &
                data$dow == 1,]

#Cut just the first 100 rows so that I'm faceting by less
Dec_2017_sundays_100 <- Dec_2017_sundays[1:100,]

Dec_2017_sun_plot <- ggplot(Dec_2017_sundays_100, aes(day,no2))+
                geom_point()+
                facet_wrap(. ~ID)
Dec_2017_sun_plot

ggsave("Dec_2017_sun_plot_non_movers.png", width = 5, height = 5)

#dataset of just Mondays in Dec 2017
Dec_2017_mondays <- data[data$year == 2017 &
                data$month == 12 &
                data$dow == 2,]

#Cut just the first 100 rows so that I'm faceting by less
Dec_2017_mondays_100 <- Dec_2017_mondays[1:100,]

Dec_2017_mon_plot <- ggplot(Dec_2017_mondays_100, aes(day,no2))+
                geom_point()+
                facet_wrap(. ~ID)
Dec_2017_mon_plot

ggsave("Dec_2017_mon_plot_non_movers.png", width = 5, height = 5)

#Tuesdays in April 2018
Apr_2018_tuesdays <- data[data$year == 2018 &
                data$month == "04" &
                data$dow == 3,]

#Cut just the first 100 rows so that I'm faceting by less
Apr_2018_tuesdays_100 <- Apr_2018_tuesdays[1:100,]

Apr_2018_tuesdays_plot <- ggplot(Apr_2018_tuesdays_100, aes(day,no2))+
                geom_point()+
                facet_wrap(. ~ID)
Apr_2018_tuesdays_plot

ggsave("April_2018_tuesdays_plot_non_movers.png", width = 5, height = 5)

#Fridays in June 2018
june_2018_fridays <- data[data$year == 2018 &
                data$month == "06" &
                data$dow == 6,]

#Cut just the first 100 rows so that I'm faceting by less
june_2018_fridays_100 <- june_2018_fridays[1:100,]

june_2018_fridays_plot <- ggplot(june_2018_fridays_100, aes(day,no2))+
                geom_point()+
                facet_wrap(. ~ID)
june_2018_fridays_plot

ggsave("june_2018_fridays_plot_non_movers.png", width = 5, height = 5)


#thursdays in nov 2019
nov_2019_thursdays <- data[data$year == 2019 &
                data$month == "11" &
                data$dow == 5,]

#Cut just the first 100 rows so that I'm faceting by less
nov_2019_thursdays_100 <- nov_2019_thursdays[1:100,]

nov_2019_thursdays_plot <- ggplot(nov_2019_thursdays_100, aes(day,no2))+
                geom_point()+
                facet_wrap(. ~ID)
nov_2019_thursdays_plot

ggsave("nov_2019_thursdays_plot_non_movers.png", width = 5, height = 5)


####
#For some individuals there is almost no variation between same weekdays of this particular month. 
#Next steps: 
#1. Restrict to people who didn't move around and look again

## Good news: comparing folks who didn't move to folks who did move postal codes in Dec 2017 Sundays and Mondays, there appears to still be a lot of variation in NO2 exposure between same days of week within a month for most participants. 

#2. Look at other days of the week 
#3. Look at 3 other sample months in the 3 years.
#4. All looks good, there is variation for most individuals. 


```

## O3

Easily a 10 ppb IQR for most individuals

```{r}

table_o3 <- do.call("rbind",tapply(data$o3, data$random_uid, quantile))
table_o3 <- as.data.frame(table_o3)
table_o3$individual <- 1:1000

quantile(table_o3$`25%`)
# 0%      25%      50%      75%     100% 
#13.70000 20.48938 21.13000 23.65437 28.23000 
 
quantile(table_o3$`75%`)
 #   0%     25%     50%     75%    100% 
#25.3000 32.3675 33.5575 35.5850 39.9375 

o3 <- ggplot(table_o3, aes(individual, `50%`, ymin = `25%`, ymax = `75%`)) +
            geom_pointrange() +
            labs( x = "Individual", y = expression("Range of O"[3]* " IQRs"), subtitle =  expression("O"[3]* " ranges of a random subset of 1000 participants"))
o3




```

## PM2.5

```{r}

table_pm25 <- do.call("rbind",tapply(data$pm25, data$random_uid, quantile))
table_pm25 <- as.data.frame(table_pm25)
table_pm25$individual <- 1:1000

quantile(table_pm25$`25%`)
#  0%     25%     50%     75%    100% 
#0.18050 2.15125 3.22975 3.55625 7.33100 

quantile(table_pm25$`75%`)
#     0%      25%      50%      75%     100% 
 #0.86100  7.77875  9.83775 11.56044 18.03050 

pm25 <- ggplot(table_pm25, aes(individual, `50%`, ymin = `25%`, ymax = `75%`)) +
            geom_pointrange() +
            labs( x = "Individual", y = expression("Range of PM"[2.5]* " IQRs"), subtitle =  expression("PM"[2.5]* " ranges of a random subset of 1000 participants"))
pm25




```

## SO2

The IQRs of sulphur dioxide are very narrow, with half of 25 % at 0.58 ppb and half of 75 % at 1.87 ppb This is narrower than the range expected to impact migraine based on previous research (Portt et al 2023) Therefore, SO2 was not included as a potential exposure effecting migraine in participants in Ontario.

```{r}

table_so2 <- do.call("rbind",tapply(data$so2, data$random_uid, quantile))
table_so2 <- as.data.frame(table_so2)
table_so2$individual <- 1:1000

quantile(table_so2$`25%`)
 # 0%       25%       50%       75%      100% 
#0.0107000 0.3017250 0.5856750 0.8281437 5.5793500 


quantile(table_so2$`75%`)
#      0%       25%       50%       75%      100% 
# 0.160300  1.091531  1.869625  2.814162 16.119350

so2 <- ggplot(table_so2, aes(individual, `50%`, ymin = `25%`, ymax = `75%`)) +
            geom_pointrange() +
            labs( x = "Individual", y = expression("Range of SO"[2]* " IQRs"), subtitle =  expression("SO"[2]* " ranges of a random subset of 1000 participants"))
so2


```
